Refactor sentry intercentop configuration
