+++ src/App/pageViews/pageViews.resolver.ts
@@ -1,9 +1,6 @@
-import { UseInterceptors } from '@nestjs/common'
 import { Args, Context, Mutation, Resolver } from '@nestjs/graphql'
-import SentryInterceptor from '../../sentry/security.interceptor'
 import PageViewsService from './pageViews.service'
 
-@UseInterceptors(SentryInterceptor)
 @Resolver(() => Number)
 class PageViewsResolver {
   constructor(private pageViewsService: PageViewsService) {}
