+++ src/App/tag.resolver.ts
@@ -8,13 +8,11 @@ import {
   Resolver,
 } from '@nestjs/graphql'
 import { Connection } from 'typeorm'
-import { UseInterceptors } from '@nestjs/common'
 import { Validate } from 'class-validator'
 import Tag from '../Database/Entities/tag.entity'
 import PaginationArgs from './pagination.args'
 import Wiki from '../Database/Entities/wiki.entity'
 import { ITag } from '../Database/Entities/types/ITag'
-import SentryInterceptor from '../sentry/security.interceptor'
 import ValidStringParams from './utils/customValidator'
 
 @ArgsType()
@@ -24,7 +22,6 @@ class TagIDArgs extends PaginationArgs {
   id!: string
 }
 
-@UseInterceptors(SentryInterceptor)
 @Resolver(() => Tag)
 class TagResolver {
   constructor(private connection: Connection) {}
