+++ src/App/tokenStats/tokenStats.resolver.ts
@@ -1,10 +1,7 @@
 import { Args, Query, Resolver } from '@nestjs/graphql'
-import { UseInterceptors } from '@nestjs/common'
 import TokenStatsService from './tokenStats.service'
 import TokenData from './models/tokenData.model'
-import SentryInterceptor from '../../sentry/security.interceptor'
 
-@UseInterceptors(SentryInterceptor)
 @Resolver()
 class TokenStatsResolver {
   constructor(private readonly tokenStatsService: TokenStatsService) {}
