+++ src/Indexer/Provider/graph.service.ts
@@ -1,8 +1,7 @@
 /* eslint-disable no-console */
-import { Injectable, UseInterceptors } from '@nestjs/common'
+import { Injectable } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { request, gql } from 'graphql-request'
-import SentryInterceptor from '../../sentry/security.interceptor'
 
 export type Hash = {
   id: string
@@ -26,7 +25,6 @@ const query = gql`
   }
 `
 
-@UseInterceptors(SentryInterceptor)
 @Injectable()
 class GraphProviderService {
   constructor(private configService: ConfigService) {}
@@ -38,7 +36,7 @@ class GraphProviderService {
     try {
       response = await request(reqUrl, query, { unixtime })
     } catch (err: any) {
-      console.error('GRAPH ERROR', err.message)
+      console.error('GRAPH ERROR', JSON.stringify(err, null, 2))
     }
     return response.ipfshashs.filter((hash: Hash) => hash.id.length === 46)
   }
