+++ src/Indexer/indexer.module.ts
@@ -1,5 +1,7 @@
 import { CacheModule, Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
+import { APP_INTERCEPTOR } from '@nestjs/core'
+import { SentryInterceptor } from '@ntegral/nestjs-sentry'
 import RunCommand from './run.command'
 import GraphProviderService from './Provider/graph.service'
 import HistoryProviderService from './Provider/history.service'
@@ -10,6 +12,7 @@ import DatabaseModule from '../Database/database.module'
 import httpModule from '../httpModule'
 import MetadataChangesService from './Store/metadataChanges.service'
 import { RevalidatePageService } from '../App/revalidatePage/revalidatePage.service'
+import SentryMod from '../sentry/sentry.module'
 
 @Module({
   imports: [
@@ -19,6 +22,7 @@ import { RevalidatePageService } from '../App/revalidatePage/revalidatePage.serv
     DatabaseModule,
     httpModule(20000),
     CacheModule.register({ ttl: 3600 }),
+    SentryMod
   ],
   controllers: [],
   providers: [
@@ -30,6 +34,10 @@ import { RevalidatePageService } from '../App/revalidatePage/revalidatePage.serv
     IPFSGetterService,
     RunCommand,
     RevalidatePageService,
+    {
+      provide: APP_INTERCEPTOR,
+      useFactory: () => new SentryInterceptor(),
+    },
   ],
 })
 class IndexerModule {}
