+++ src/Relayer/resolvers/relayer.resolver.ts
@@ -1,12 +1,9 @@
-import { UseInterceptors } from '@nestjs/common'
 import { Args, Mutation, Resolver } from '@nestjs/graphql'
-import SentryInterceptor from '../../sentry/security.interceptor'
 
 import RelayerService from '../services/relayer.service'
 import SignaturePayloadInput from './dto/signaturePayload.dto'
 import Relayer from './models/relayer'
 
-@UseInterceptors(SentryInterceptor)
 @Resolver()
 class RelayerResolver {
   constructor(private relayerService: RelayerService) {}
