--- src/sentry/security.interceptor.ts
@@ -1,22 +1,0 @@
-import {
-  ExecutionContext,
-  Injectable,
-  NestInterceptor,
-  CallHandler,
-} from '@nestjs/common'
-import { Observable } from 'rxjs'
-import { tap } from 'rxjs/operators'
-import * as Sentry from '@sentry/minimal'
-
-@Injectable()
-class SentryInterceptor implements NestInterceptor {
-  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
-    return next.handle().pipe(
-      tap(null, exception => {
-        Sentry.captureException(exception)
-      }),
-    )
-  }
-}
-
-export default SentryInterceptor
